v_2/v2_requirements.txt
Babel==2.14.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask-Babel==2.0.0
Flask-Cors==4.0.0
greenlet==3.0.3
h3==3.7.7
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.1
Markdown==3.5
MarkupSafe==2.1.3
mysql-connector==2.2.9
packaging==23.2
parameterized==0.9.0
pymongo==4.6.2
pytz==2024.1
redis==5.0.3
requests==2.31.0
setuptools==68.2.2
SQLAlchemy==2.0.29
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.2
